% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arquivos.r
\name{classifica_uri}
\alias{classifica_uri}
\title{Classifica Um Caminho De Acordo Com Fonte}
\usage{
classifica_uri(uri)
}
\arguments{
\item{uri}{caminho de dados a classificar}
}
\value{
uri com classe \code{"uri_s3"} ou \code{"uri_local"}
}
\description{
Classifica um caminho de dados de acordo com ser local ou bucket no s3
}
