% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.r
\name{coltype2dbinterface}
\alias{coltype2dbinterface}
\title{Converte Tipo De Coluna Para Padrao \code{dbinterface}}
\usage{
coltype2dbinterface(schema)
}
\arguments{
\item{schema}{um schema de tabela como retornado por \code{\link{get_schemas}}}
}
\value{
\code{schema} com tipos de colunas convertidos para padrao do \code{dbinterface}
}
\description{
Auxiliar para converter chaves tipo de coluna para os nomes reconhecidos pelo \code{dbinterface}
}
