% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conector-pfv.r
\name{conectamock_pfv}
\alias{conectamock_pfv}
\title{Construtor Externo De Conexao Com Banco}
\usage{
conectamock_pfv(uri)
}
\arguments{
\item{uri}{caminho dos dados, local ou bucket no s3; neste segundo caso informar bucket e prefixo
como um caminho completo, em string}
}
\value{
objeto \code{mock}; Veja \code{\link[dbrenovaveis:conectamock]{dbrenovaveis::conectamock}}
}
\description{
Gera uma conexao com banco no padrao \code{dbrenovaveis} para uso das funcionalidades de leitura
}
\details{
Usualmente o \code{dbrenovaveis} espera encontrar um arquivo \code{"schema.json"} a partir do qual montar o
banco de dados. No caso do modelo PFV, e necessario fazer interface com os dados sem este
arquivo, se baseando em schemas internamente definidos pelo pacote.

\code{conectamock_pfv} consiste em uma "volta" no processo natural de construcao do \code{mock} a partir de
um \code{"schema.json"}, realizando diversas de suas tarefas a partir dos schemas internos para gerar
entao uma conexao com o banco
}
\examples{

uri <- system.file("extdata/", package = "pfvIO")
conn <- conectamock_pfv(uri)

\dontrun{
# caso de um banco em bucket s3
uri <- "s3://nome_do_bucket/prefixo/do/banco"
conn <- conectamock_pfv(uri)
}

}
