% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{extend_numeric}
\alias{extend_numeric}
\title{Aumenta Tipo \code{"numeric"} Para Conter \code{"integer"} Tambem}
\usage{
extend_numeric(tipos)
}
\arguments{
\item{tipos}{vetor ou lista nomeado de tipos de dados}
}
\value{
\code{tipos} com posicoes originalmente \code{"numeric"} aumentadas para \code{c("numeric", "integer")}
}
\description{
Auxiliar para validacao de tipos de colunas
}
