% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.r
\name{fix_uri_filetype}
\alias{fix_uri_filetype}
\title{Adiciona Elementos \code{uri} e \code{fileType} A Um Schema}
\usage{
fix_uri_filetype(schema, uri, fileType)
}
\arguments{
\item{schema}{um schema unitario lido por \code{get_single_schema()}}

\item{uri}{string indicando a uri a substituir}

\item{fileType}{string indicando a extensao (com \code{"."} inicial) a substituir}
}
\value{
\code{schema} com elementos \code{uri} e \code{fileType} substituidos
}
\description{
Auxiliar para compatibilizar o schema fixo com partes dinamicas, dependendo da fonte de dados
}
