% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.r
\name{get_config}
\alias{get_config}
\title{Getter Do Arquivo De Configuracao}
\usage{
get_config(conn, ...)
}
\arguments{
\item{conn}{um objeto de conexao com banco como gerado por \code{\link{conectamock_pfv}}}

\item{...}{sem uso, existe apenas para consistencia com as outras getters}
}
\value{
lista do arquivo de configuracao lido
}
\description{
Wrapper para acesso ao arquivo de configuracao no caminho de dados
}
