% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.r
\name{get_dataset}
\alias{get_dataset}
\title{Getter De Dataset Completo}
\usage{
get_dataset(conn, ...)
}
\arguments{
\item{conn}{um objeto de conexao com banco como gerado por \code{\link{conectamock_pfv}}}

\item{...}{definicoes de subset. Veja \code{\link{pfvio_getters}}}
}
\value{
lista de \code{data.table}s contendo todas as tabelas do banco nos subsets definidos por \code{...}
}
\description{
Acessa todas as tabelas em uma conexao aplicando subsets igualmente
}
