% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.r
\name{get_model_artifact}
\alias{get_model_artifact}
\title{Getter Do Artefato De Modelo}
\usage{
get_model_artifact(filename, path)
}
\arguments{
\item{filename}{nome do arquivo do artefato a ser lido SEM EXTENSAO; por padrao todos os
artefatos sao salvos como objetos rds}

\item{path}{caminho onde o artefato esta salvo}
}
\value{
objeto lido do arquivo rds
}
\description{
Wrapper para acesso a artefatos de modelos gerados
}
