% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.r
\name{get_schemas}
\alias{get_schemas}
\title{Getter De Schemas De Tabelas}
\usage{
get_schemas(tabelas)
}
\arguments{
\item{tabelas}{vetor com nomes de tabelas; sera validado e retorna erro caso alguma tabela
solicitada nao conste no schema do pfv}
}
\value{
lista nomeada de schemas por tabela
}
\description{
Wrapper para buscar schemas de multiplas tabelas suportadas pelo pacote
}
