% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.r
\name{get_single_schema}
\alias{get_single_schema}
\title{Getter De Schema Unitario}
\usage{
get_single_schema(tabela)
}
\arguments{
\item{tabela}{nome da tabela cujo schema deve ser buscado}
}
\value{
lista representando schema da tabela
}
\description{
Wrapper para buscar schema de uma unica tabela
}
