% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{guessers}
\alias{guessers}
\alias{guess_col_names}
\alias{guess_col_types}
\alias{guess_col_limits}
\title{Extratores De Caracteristicas De Tabela}
\usage{
guess_col_names(tabela)

guess_col_types(tabela)

guess_col_limits(tabela)
}
\arguments{
\item{tabela}{a tabela da qual extrair informacoes}
}
\value{
nomes, tipos ou limites das colunas, dependendo da funcao chamada
}
\description{
Auxiliares para extrair de um schema os nomes, tipos e limites de cada coluna da tabela
}
