% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arquivos.r
\name{id_tipo_arquivo}
\alias{id_tipo_arquivo}
\title{Identifica Extensao De Arquivo}
\usage{
id_tipo_arquivo(arquivos)
}
\arguments{
\item{arquivos}{lista de nomes de arquivos}
}
\value{
vetor com comprimento de \code{arquivos} contendo a extensao de cada um
}
\description{
Identifica Extensao De Arquivo
}
