% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{is_in_bounds}
\alias{is_in_bounds}
\title{Checa Se Valores Em \code{x} Pertencem Ao Intervalo Fechado \code{bounds}}
\usage{
is_in_bounds(x, bounds)
}
\arguments{
\item{x}{vetor numerico}

\item{bounds}{vetor de duas posicoes indicando o intervalo de dados}
}
\value{
booleano indicando se todos os valores de \code{x} pertencem a \verb{[bounds[1], bounds[2]]}
}
\description{
Checa Se Valores Em \code{x} Pertencem Ao Intervalo Fechado \code{bounds}
}
