% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arquivos.r
\name{lista_arquivos}
\alias{lista_arquivos}
\title{Lista Arquivos Em Um Caminho}
\usage{
lista_arquivos(uri)
}
\arguments{
\item{uri}{string indicando um diretorio local ou bucket no s3}
}
\value{
vetor de arquivos existentes em \code{uri}
}
\description{
Wrapper para listar arquivos em um diretorio local ou bucket s3
}
