% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.r
\name{logger_setup}
\alias{logger_setup}
\title{Gera Um Objeto \code{lgr::Logger}}
\usage{
logger_setup(level = Sys.getenv("LOG_LEVEL", unset = "info"))
}
\arguments{
\item{level}{nivel de logging. Procura a variavel de ambiente \code{"LOG_LEVEL"} e, caso nao exista,
usa \code{"info"}}
}
\value{
objeto \code{lgr::Logger}
}
\description{
Funcao interna para uso no loading do pacote, nao deve ser chamada diretamente pelo usuario
}
