% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfvIO-package.R
\docType{package}
\name{pfvIO-package}
\alias{pfvIO}
\alias{pfvIO-package}
\title{pfvIO: Funcionalidades De Input/Output Para O Modelo De Previsão De Geracao Solar Fotovoltaica}
\description{
Este e um pacote companion do modelo de previsao de geracao solar fotovoltaica. Contem wrappers de validacao, leitura e escrita dos dados associados a execucao deste modelo.
}
\author{
\strong{Maintainer}: Lucas Khenayfis \email{lkhenayfis@ons.org.br}

}
\keyword{internal}
