% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.r
\name{pfvio_getters}
\alias{pfvio_getters}
\alias{get_usinas}
\alias{get_potencia_disponivel_observada}
\alias{get_geracao_observada}
\alias{get_corte_observado}
\alias{get_irradiancia_prevista}
\alias{get_melhor_historico_geracao}
\alias{get_melhor_historico_geracao_sem_cortes}
\title{Getters De Dados Individuais}
\usage{
get_usinas(conn, ...)

get_potencia_disponivel_observada(conn, ...)

get_geracao_observada(conn, ...)

get_corte_observado(conn, ...)

get_irradiancia_prevista(conn, ...)

get_melhor_historico_geracao(conn, ...)

get_melhor_historico_geracao_sem_cortes(conn, ...)
}
\arguments{
\item{conn}{um objeto de conexao com banco como gerado por \code{\link{conectamock_pfv}}}

\item{...}{definicoes de subset. Veja Detalhes e Exemplos}
}
\value{
todas as funcoes retornam um \code{data.table}, de dado que depende de qual funcao foi
chamada, com subsets em \code{...} aplicados
}
\description{
Wrappers para acesso as tabelas do PFV em uma determinada conexao
}
\details{
O argumento \code{...} permite passar argumentos homonimos de colunas da tabela, cujos valores serao
utilizados para realizar subset do dado.

Se a coluna e de string ou inteiros, o subset sera realizado buscando onde aquela coluna assume o
valor passado. Por exemplo, para um dado com uma coluna \code{classe}, passar o argumento
\code{classe = "nivel_1"} realizaria o subset de todas as linhas nas quais a coluna \code{classe} tem valor
\code{"nivel_1"}.

No caso de colunas de data ou datahora, existe uma implementacao especial para facilitar a
especificacao de faixas temporais. Nestes casos pode ser passada uma string no formato
\code{"YYYY[-MM-[DD[ HH[:MM[:SS]]]]]/YYYY[-MM-[DD[ HH[:MM[:SS]]]]]"}. As partes entre colchetes sao
opcionais, tanto no limite inicial quanto final. Ademais, pode ser passada uma string com apenas
um dos limites; neste caso sera interpretado como "tudo a partir de" (caso de somente limite
inicial) ou "tudo ate" (caso contrario).
}
\examples{

# conexao com o banco
dir  <- system.file("extdata/", package = "pfvIO")
conn <- conectamock_pfv(dir)

# buscando geracao observada de uma usina especifica e todas as fontes
dt <- get_geracao_observada(conn, id_usina = "BAUFI1")

# mesma usina e apenas dados do PI
dt <- get_geracao_observada(conn, id_usina = "BAUFI1", id_fonte_observacao = "PI")

# conjunto de tres usinas e apenas dados do PI
dt <- get_melhor_historico_geracao(conn, id_usina = c("BAUFI1", "BAUFI2", "BAUFI3"))

# EXEMPLOS DE SUBSET DE DATA ---------------------------

# pegando apenas o dia de 2025-06-10
dt <- get_geracao_observada(conn, id_usina = "BAUFI1", data_hora_observacao = "2025-06-10")

# pegando janela 2025-06-20 10:00:00 a 2025-06-25 19:30:00
dt <- get_geracao_observada(conn, id_usina = "BAUFI1",
    data_hora_observacao = "2025-06-20 10:00:00/2025-06-25 19:30:00")

}
\seealso{
\code{\link{get_config}}
}
