% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arquivos.r
\name{split_bucket_prefix}
\alias{split_bucket_prefix}
\title{Helper Para Parse De Uri s3}
\usage{
split_bucket_prefix(uri)
}
\arguments{
\item{uri}{caminho completo no s3}
}
\value{
vetor de duas posicoes: bucket e prefixo separados de \code{uri}
}
\description{
Separa uma uri s3 em dois elementos: bucket e prefixo
}
