% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arquivos.r
\name{supported_files}
\alias{supported_files}
\title{Wrapper De Lista De Tipos De Arquivos Suportados}
\usage{
supported_files(as_regex = FALSE)
}
\arguments{
\item{as_regex}{booleano indicando se extensoes devem ser retornadas como padrao para regex}
}
\value{
se \code{as_regex = FALSE} (padrao), vetor de extensoes; do contrario uma string combinando
todas em padrao para uso com funcoes de regex (\code{sub}, \code{grep} e etc.)
}
\description{
Funcao dummy que retorna vetor de extensoes suportadas, opcionalmente como padrao para regex
}
\details{
Atualmente sao suportados os seguintes tipos de arquivos tabulares
\itemize{
\item \code{.csv}
\item \code{.parquet}
\item \code{.parquet.gzip}
}

... e arquivos de lista chave-valor
\itemize{
\item \code{.json}
\item \code{.jsonc}
}
}
