% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{supported_tables}
\alias{supported_tables}
\title{Wrapper De Lista De Tabelas Suportadas}
\usage{
supported_tables()
}
\value{
vetor de nomes das tabelas suportadas no pacote
}
\description{
Funcao dummy que retorna vetor de tabelas suportadas
}
\details{
Atualmente sao suportados os seguintes tipos de arquivos tabulares
\itemize{
\item \code{usinas}
\item \code{potencia_disponivel_observada}
\item \code{geracao_observada}
\item \code{corte_observado}
\item \code{irradiancia_prevista}
\item \code{melhor_historico_geracao}
\item \code{melhor_historico_geracao_sem_cortes}
}
}
