% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{valida_filtra_tabelas}
\alias{valida_filtra_tabelas}
\title{Validador De Nomes De Tabelas Do Banco}
\usage{
valida_filtra_tabelas(tabelas)
}
\arguments{
\item{tabelas}{um vetor de nomes de tabelas a serem batidos contra as existentes no banco}
}
\value{
\code{NULL} invisivel, se \code{tabelas} nao contiver nomes invalidos; levanta erro do contrario
}
\description{
Checa se os nomes em \code{tabelas} sao nomes validos no schema do banco
}
\details{
O PFV atualmente suporta as tabelas
\itemize{
\item \code{usinas}
\item \code{potencia_disponivel_observada}
\item \code{geracao_observada}
\item \code{corte_observado}
\item \code{irradiancia_prevista}
\item \code{melhor_historico_geracao}
\item \code{melhor_historico_geracao_sem_cortes}
}

Qualquer nome diferente destes levantara um erro
}
