% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{valida_limites_colunas}
\alias{valida_limites_colunas}
\title{Validador De Limites Numericos}
\usage{
valida_limites_colunas(dt, limites)
}
\arguments{
\item{dt}{um \code{data.table} no qual realizar checagem}

\item{limites}{lista nomeada de vetores numericos. Veja Detalhes}
}
\value{
\code{NULL} se todos as checagens forem aprovadas; levanta erro do contrario
}
\description{
Checa se colunas do dado \code{dt} pertencem aos intervalos declarados em \code{limites}
}
\details{
\code{limites} deve ser uma lista nomeada de vetores de duas posicoes cada. Cada vetor representa os
limites de valores da coluna correspondente a seu nome na lista. So serao checadas as colunas
com elementos homonimos em \code{limites}
}
