% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{valida_nomes_colunas}
\alias{valida_nomes_colunas}
\title{Validador De Nomes De Colunas}
\usage{
valida_nomes_colunas(dt, nomes)
}
\arguments{
\item{dt}{\code{data.table} no qual checar nomes}

\item{nomes}{vetor de nomes de colunas para procurar em \code{dt}}
}
\value{
\code{NULL} se todos as checagens forem aprovadas; levanta erro do contrario
}
\description{
Checa se colunas em \code{nomes} constam no dado \code{dt}
}
