% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.r
\name{valida_tipos_colunas}
\alias{valida_tipos_colunas}
\title{Validador De Tipos De Colunas}
\usage{
valida_tipos_colunas(dt, tipos)
}
\arguments{
\item{dt}{um \code{data.table} no qual realizar checagem}

\item{tipos}{lista nomeada de vetores numericos. Veja Detalhes}
}
\value{
\code{NULL} se todos as checagens forem aprovadas; levanta erro do contrario
}
\description{
Checa se colunas do dado \code{dt} tem tipos declarados em \code{tipos}
}
\details{
\code{tipos} deve ser uma lista nomeada de tipos de dados (i.e. \code{"Date"}, \code{"numeric"} e etc.). Cada
elemento representa os tipos de valores da coluna correspondente a seu nome na lista. So serao
checadas as colunas com elementos homonimos em \code{tipos}
}
