% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.r
\name{write_dataset}
\alias{write_dataset}
\title{Escritor De Dados Tabelados}
\usage{
write_dataset(object, filename, path)
}
\arguments{
\item{object}{objeto a ser salvo}

\item{filename}{nome do arquivo do artefato a ser lido INCLUINDO EXTENSAO; observe que apenas
\code{c("csv", "parquet", "parquet.gzip")} sao suportados atualmente}

\item{path}{caminho onde o artefato esta salvo}
}
\value{
nao ha retorno; apenas salva \code{object} como \code{path/filename.rds}
}
\description{
Wrapper para salvamento de dados tabelados em formatos suportados
}
