% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escrita.r
\name{write_model_artifact}
\alias{write_model_artifact}
\title{Escritor De Artefato De Modelo}
\usage{
write_model_artifact(object, filename, path)
}
\arguments{
\item{object}{objeto a ser salvo}

\item{filename}{nome do arquivo do artefato a ser lido SEM EXTENSAO; por padrao todos os
artefatos sao salvos como objetos rds}

\item{path}{caminho onde o artefato esta salvo}
}
\value{
nao ha retorno; apenas salva \code{object} como \code{path/filename.rds}
}
\description{
Wrapper para salvamento de artefatos de modelos gerados
}
