% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.r
\name{coltype2dbrenovaveis}
\alias{coltype2dbrenovaveis}
\title{Converte Tipo De Coluna Para Padrao \code{dbrenovaveis}}
\usage{
coltype2dbrenovaveis(schema)
}
\description{
Auxiliar para converter chaves tipo de coluna para os nomes reconhecidos pelo \code{dbrenovaveis}
}
