% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conector-pfv.r
\name{fake_mock}
\alias{fake_mock}
\title{Construtor Interno De Conexao Com Banco}
\usage{
fake_mock(tabelas, uri)
}
\arguments{
\item{tabelas}{lista de tabelas preprocessadas}

\item{uri}{caminho local ou no s3 do banco. Veja \code{\link{conectamock_pfv}}}
}
\value{
objeto \code{mock} de conexao com o banco
}
\description{
Gera um objeto \code{mock} "por fora" do \code{dbrenovaveis}
}
