% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.r
\name{get_pkg_logger}
\alias{get_pkg_logger}
\title{Getter Do Logger Interno Do Pacote}
\usage{
get_pkg_logger()
}
\value{
objeto \code{lgr::Logger} gerado por \code{logger_setup()}
}
\description{
Funcao auxiliar para acesso ao logger do pacote
}
