% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.r
\name{getters_schemas_individuais}
\alias{getters_schemas_individuais}
\alias{get_schema_usinas}
\alias{get_schema_potencia_disponivel_observada}
\alias{get_schema_geracao_observada}
\alias{get_schema_corte_observado}
\alias{get_schema_irradiancia_prevista}
\alias{get_schema_melhor_historico_geracao}
\title{Wrappers De Schemas Individuais}
\usage{
get_schema_usinas()

get_schema_potencia_disponivel_observada()

get_schema_geracao_observada()

get_schema_corte_observado()

get_schema_irradiancia_prevista()

get_schema_melhor_historico_geracao()
}
\value{
lista representando o schema da tabela
}
\description{
Funcoes dummy que retornam um schema para cada tabela suportada pelo pfv
}
